Allow not writing id and keys
This skips writing a device id and keys (still writing bootloader and
calibration), preparing for a new system where a device is registered
after it is assembled instead of registering chips early.

In this mode, if there is an existing EEPROM magic value, the entire
EEPROM is cleared (this can be used to reset chips later to re-register
them).

This mode is now also the default, you can change back to writing an id
by entering an id to write (from them on the old mode of sequential ids
is entered again and can be left by resetting the programmer).